small adjustments mainly to absence requests and work time model schedule

+++ database/migrations/2025_08_06_100001_create_work_time_model_schedules_table.php
@@ -20,7 +20,7 @@ return new class() extends Migration
             $table->unsignedTinyInteger('weekday');
             $table->time('start_time')->nullable();
             $table->time('end_time')->nullable();
-            $table->decimal('break_minutes', 5, 2)->default(0);
+            $table->unsignedInteger('break_minutes')->default(0);
             $table->decimal('work_hours', 4, 2)->default(0);
 
             $table->timestamp('created_at')->nullable();
